requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
crypto==1.4.1
cryptography==42.0.8
cssselect2==0.7.0
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
firebase-admin==6.5.0
fonttools==4.53.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
Naked==0.1.32
openpyxl==3.1.4
pillow==10.3.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydyf==0.10.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyphen==0.15.0
PyYAML==6.0.1
razorpay==1.4.2
requests==2.32.3
rsa==4.9
service-identity==24.1.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.3.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==62.2
webencodings==0.5.1
zope.interface==6.4.post2
zopfli==0.2.3
